ranking different errors, might need to revisit

+++ error.py
@@ -5,7 +5,7 @@ saved_errors = []
 
 def init_saved_errors():
     for i in range(20):
-        saved_errors.append((0, set()))
+        saved_errors.append((0, set(), set()))
 
 def unseen_errors(error_input):
     return [error for error in split_errors(error_input) if categorise_error(error) == set() and error.strip() != ""]
@@ -18,6 +18,7 @@ def update_saved_errors(error_input:str, tmp_input_file_location: str):
     heap_buffer_overflow_priority = 1
     current_priority = 0
     type_priority = 50
+    error_pos = get_errors_with_filepos(error_input)
     for error in current_type:
         if error == ErrorType.SIG_SEGV:
             current_priority += segv_priority
@@ -28,7 +29,12 @@ def update_saved_errors(error_input:str, tmp_input_file_location: str):
     lowest_priority_idx = 21
     min_priority = 9999
     new_type = True
+    already_stored = False
     for i in range(20):
+        stored_error_pos = saved_errors[i][2]
+        if not stored_error_pos.symmetric_difference(error_pos):
+            already_stored = True
+            break
         stored_type = saved_errors[i][1]
         priority = saved_errors[i][0]
         if priority < min_priority:
@@ -36,10 +42,13 @@ def update_saved_errors(error_input:str, tmp_input_file_location: str):
             lowest_priority_idx = i
         if not stored_type.symmetric_difference(current_type):
             new_type = False
+    if already_stored:
+        print(f"already stored the same error!")
+        return set()
     if new_type:
         current_priority += 1000
 
     if current_priority > min_priority and lowest_priority_idx != 21:
-        saved_errors[lowest_priority_idx] = (current_priority, current_type)
+        saved_errors[lowest_priority_idx] = (current_priority, current_type, error_pos)
         shutil.copy(tmp_input_file_location, f"fuzzed-tests/input_{lowest_priority_idx}.cnf")
     return current_type
